requirements.txt
absl-py==2.3.0
annotated-types==0.7.0
anyio==4.9.0
azure-core==1.34.0
azure-storage-blob==12.25.1
bert-score==0.3.13
blinker==1.9.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
contourpy==1.3.2
cryptography==45.0.4
cycler==0.12.1
distro==1.9.0
docopt==0.6.2
faiss-cpu==1.11.0
filelock==3.18.0
Flask==3.1.1
flask-cors==6.0.1
fonttools==4.58.4
fsspec==2025.5.1
fuzzywuzzy==0.18.0
gunicorn==23.0.0
h11==0.16.0
hf-xet==1.1.3
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.33.0
idna==3.10
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
kiwisolver==1.4.8
loguru==0.7.3
MarkupSafe==3.0.2
matplotlib==3.10.3
mpmath==1.3.0
networkx==3.5
nltk==3.9.1
numpy==2.3.0
openai==1.86.0
packaging==25.0
pandas==2.3.0
pillow==11.2.1
pycparser==2.22
pydantic==2.11.6
pydantic_core==2.33.2
PyMuPDF==1.26.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
rouge_score==0.1.2
safetensors==0.5.3
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
sympy==1.14.0
tiktoken==0.9.0
tokenizers==0.21.1
torch==2.7.1
tqdm==4.67.1
transformers==4.52.4
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
Werkzeug==3.1.3
yarg==0.1.10
